add reentrant_lock_op in main_job

+++ oneflow/core/job/runtime.cpp
@@ -79,9 +79,9 @@ Runtime::Runtime(const Plan& plan, size_t total_piece_num, bool is_experiment_ph
   // if is this machine master
   FOR_RANGE(int64_t, i, 0,
             Global<std::vector<std::unique_ptr<JobDesc>>>::Get()->at(0)->TotalBatchNum()) {
-    Global<BufferMgr<int32_t>>::Get()->Get(kChannelNameGlobalWaitJobId)->Send(0);
+    Global<BufferMgr<int64_t>>::Get()->Get(kChannelNameGlobalWaitJobId)->Send(0);
   }
-  Global<BufferMgr<int32_t>>::Get()->Get(kChannelNameGlobalWaitJobId)->Close();
+  Global<BufferMgr<int64_t>>::Get()->Get(kChannelNameGlobalWaitJobId)->Close();
   runtime_ctx->WaitUntilCntEqualZero("running_actor_cnt");
   OF_BARRIER();
   DeleteAllGlobal();
